+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for aHaskellPriori
 
+## 2019.10.26.1
+- Main.hs aangepast, begin gemaakt aan het menu
+- Variabelen voor debuggen in `DatatypesOld.hs` gezet
+- `Datatypes.hs` bevatten de nieuwe datatypes `Product`, `Transactie`, `Transacties`. Helaas nog niet werkend
+
 ## 2019.10.24.1
 - Documentatie wordt in de goede map aangemaakt door `stack haddock --haddock-arguments "-o docs"`